Peer review: making claim result shareable for all the strategies
