+++ src/main/java/com/sonyericsson/hudson/plugins/gerrit/trigger/coordination/hazelcast/HazelcastNotificationClaimStrategy.java
@@ -23,6 +23,8 @@ package com.sonyericsson.hudson.plugins.gerrit.trigger.coordination.hazelcast;
 
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.map.IMap;
+import com.sonyericsson.hudson.plugins.gerrit.trigger.spi.ClaimResult;
+import com.sonyericsson.hudson.plugins.gerrit.trigger.spi.ClaimResults;
 import com.sonyericsson.hudson.plugins.gerrit.trigger.spi.NotificationClaimStrategy;
 import com.sonymobile.tools.gerrit.gerritevents.dto.events.GerritTriggeredEvent;
 import edu.umd.cs.findbugs.annotations.NonNull;
@@ -30,7 +32,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.concurrent.TimeUnit;
-import java.util.function.Consumer;
 
 /**
  * Hazelcast-backed implementation of NotificationClaimStrategy for HA/HS deployments.
@@ -103,9 +104,9 @@ public class HazelcastNotificationClaimStrategy extends NotificationClaimStrateg
             // Fail-open: execute the notification action even without claiming
             try {
                 claimed.run();
-                return new SuccessfulClaim();
+                return ClaimResults.success();
             } catch (Exception e) {
-                return new FailedClaim(e);
+                return ClaimResults.failed(e);
             }
         }
 
@@ -127,16 +128,16 @@ public class HazelcastNotificationClaimStrategy extends NotificationClaimStrateg
                 logger.debug("Successfully claimed notification right for event: {}", eventId);
                 try {
                     claimed.run();
-                    return new SuccessfulClaim();
+                    return ClaimResults.success();
                 } catch (Exception actionException) {
                     logger.error("Error executing notification action after successful claim: {}", eventId,
                             actionException);
-                    return new FailedClaim(actionException);
+                    return ClaimResults.failed(actionException);
                 }
             } else {
                 // Another replica already claimed notification
                 logger.debug("Another replica already claimed notification for event: {}", eventId);
-                return new NotClaimedResult();
+                return ClaimResults.notClaimed();
             }
         } catch (Exception e) {
             // Hazelcast operation failed
@@ -146,9 +147,9 @@ public class HazelcastNotificationClaimStrategy extends NotificationClaimStrateg
                 claimed.run();
             } catch (Exception innerException) {
                 logger.error("Error executing notification action after claim failure", innerException);
-                return new FailedClaim(innerException);
+                return ClaimResults.failed(innerException);
             }
-            return new SuccessfulClaim();
+            return ClaimResults.success();
         }
     }
 
@@ -177,83 +178,5 @@ public class HazelcastNotificationClaimStrategy extends NotificationClaimStrateg
             }
         }
         return DEFAULT_NOTIFICATION_CLAIM_TTL_MINUTES;
-    }
-
-    /**
-     * Successful claim result - the action was executed.
-     */
-    private static class SuccessfulClaim implements ClaimResult {
-        @Override
-        @NonNull
-        public ClaimResult notClaimed(@NonNull Runnable notClaimed) {
-            // Claim was successful, don't run notClaimed handler
-            return this;
-        }
-
-        @Override
-        @NonNull
-        public ClaimResult onError(@NonNull Consumer<Exception> onError) {
-            // No error occurred
-            return this;
-        }
-    }
-
-    /**
-     * Not claimed result - another instance already claimed the notification.
-     */
-    private static class NotClaimedResult implements ClaimResult {
-        @Override
-        @NonNull
-        public ClaimResult notClaimed(@NonNull Runnable notClaimed) {
-            // Run the notClaimed handler
-            try {
-                notClaimed.run();
-            } catch (Exception e) {
-                logger.error("Error in notClaimed handler", e);
-            }
-            return this;
-        }
-
-        @Override
-        @NonNull
-        public ClaimResult onError(@NonNull Consumer<Exception> onError) {
-            // No error occurred (just not claimed)
-            return this;
-        }
-    }
-
-    /**
-     * Failed claim result - an error occurred during processing.
-     */
-    private static class FailedClaim implements ClaimResult {
-        private final Exception exception;
-
-        /**
-         * Constructor.
-         *
-         * @param exception the exception that occurred
-         */
-        FailedClaim(Exception exception) {
-            this.exception = exception;
-        }
-
-        @Override
-        @NonNull
-        public ClaimResult notClaimed(@NonNull Runnable notClaimed) {
-            // Don't run notClaimed - this was an error, not a "not claimed" situation
-            return this;
-        }
-
-        @Override
-        @NonNull
-        public ClaimResult onError(@NonNull Consumer<Exception> onError) {
-            // Run the error handler
-            try {
-                onError.accept(exception);
-            } catch (Exception e) {
-                logger.error("Error in error handler", e);
-            }
-            return this;
-        }
     }
 }
